Artix_Urbinx/Artix_UrbanX_API/requirements.txt
fastapi
uvicorn
ultralytics
supervision
opencv-python
geopandas
shapely
pyproj
Pillow